requirements.txt
fastapi==0.115.5
pillow==11.0.0
numpy==2.1.3
